Core/DataStores: 9.1.0 db2 structures
Port From (https://github.com/TrinityCore/TrinityCore/commit/ce1046a8fd0f04aad6c414786949269fffea88f3)

+++ Source/Game/Entities/Item/ItemTemplate.cs
@@ -307,7 +307,7 @@ namespace Game.Entities
         public HolidayIds GetHolidayID() { return (HolidayIds)ExtendedData.RequiredHoliday; }
         public float GetDmgVariance() { return ExtendedData.DmgVariance; }
         public byte GetArtifactID() { return ExtendedData.ArtifactID; }
-        public byte GetRequiredExpansion() { return ExtendedData.ExpansionID; }
+        public byte GetRequiredExpansion() { return (byte)ExtendedData.ExpansionID; }
 
         public bool IsCurrencyToken() { return (GetBagFamily() & BagFamilyMask.CurrencyTokens) != 0; }
